Create a read-only DataFrame view
