+++ d3x-morpheus-core/src/main/java/com/d3x/morpheus/matrix/D3xMatrixView.java
@@ -241,9 +241,7 @@ public interface D3xMatrixView {
      *
      * @return {@code true} iff this matrix contains no elements.
      */
-    default boolean isEmpty() {
-        return size() == 0;
-    }
+    boolean isEmpty();
 
     /**
      * Identifies square matrices.
